+++ packages/isar/lib/src/web/isar_collection_impl.dart
@@ -176,7 +176,10 @@ class IsarCollectionImpl<OBJ> extends IsarCollection<OBJ> {
       unsupportedOnWeb();
 
   @override
-  int getSizeSync({bool includeIndexes = false, bool includeLinks = false}) =>
+  int getSizeSync({
+    bool includeIndexes = false,
+    bool includeLinks = false,
+  }) =>
       unsupportedOnWeb();
 
   @override